pyproject.toml
[tool.poetry]
name = "skimapper"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
gpxpy = "^1.5.0"
stravaio = "^0.0.9"
matplotlib = "^3.6.2"
PyYAML = "^6.0"
requests = "^2.28.1"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
